Changed length parameter of Incremental.Kick.Helpers.PasswordGenerator.Generate from int32 to byte due to failures of unit tests. int32 allowed negative values (which resulted in errors in code) and int.MaxValue resulted in outofmemory exception. web site currently uses only 8 characters for password so lengths of 0-255 should be acceptable. Also renamed PasswordGeneratorTest to Tests to fit other test files.

+++ trunk/DotNetKicks/Incremental.Kick/Helpers/PasswordGenerator.cs
@@ -4,7 +4,7 @@ using System.Text;
 
 namespace Incremental.Kick.Helpers {
     public class PasswordGenerator {
-        public static string Generate(int passwordLength) {
+        public static string Generate(byte passwordLength) {
             char[] vowels = new char[] { 'a', 'e', 'i', 'o', 'u' };
             char[] consonants = new char[] { 'b', 'c', 'd', 'f', 'g', 'h', 'j', 'k', 'l', 'm', 'n', 'p', 'r', 's', 't', 'v' };
             char[] doubleConsonants = new char[] { 'c', 'd', 'f', 'g', 'l', 'm', 'n', 'p', 'r', 's', 't' };
